requirements.txt
requests==2.32.3
python-telegram-bot==20.0
APScheduler==3.10.4
python-dotenv==1.0.1
